refactor: extract ranking helpers into search-ranking module
